requirements.txt
numpy
pandas
scikit-learn
xgboost
torch
gymnasium
stable-baselines3
ta
transformers
pydantic
pyyaml
ccxt
scipy
fastapi
uvicorn
pytest
pytest-asyncio
torch-geometric
ccxt
scipy
fastapi
uvicorn
pytest
pytest-asyncio